+++ testdata/next/app/(marketing)/about/page.tsx
@@ -1,3 +1,10 @@
+import { Chat } from "@/components"
+
 export default function About() {
-  return <p>About us</p>
+  return (
+    <div>
+      <p>About us</p>
+      <Chat />
+    </div>
+  )
 }