+++ testdata/next/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Shell } from "./Shell"
+export { default as Chat } from "./Chat"
+export { default as DataTable } from "./DataTable"